I really really don't want to change

+++ src/main/java/com/halcyon/online_store/entity/Product.java
@@ -27,7 +27,7 @@ import lombok.EqualsAndHashCode;
 
     private Long tid;
 
-    private Long pid;
+    private int pid;
 
 
 }
